backend/requirements.txt
fastapi
paho-mqtt
python-dotenv
sqlalchemy
uvicorn[standard]
jinja2
httpx
